Refactored sql queries into a single file
